# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get a list of courses",
                "operationId": "d452dd5ed2fa44f48bd98e5677e3cc3f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Store a new course",
                "operationId": "661e120d4c0deb68ae037ccc30104d02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Web Development 101"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "An introductory course to web development"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course created"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get course details",
                "operationId": "bd82915c39d2443dd443544603b03388",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "Update a course",
                "operationId": "6ba0bbb7b5ffadc1fee38b252c41cde8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Advanced Web Development"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Learn advanced web development techniques"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "Delete a course",
                "operationId": "fdb4bda845d6477a21a30d102dd029b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/subcategories/{parentId}": {
            "get": {
                "tags": [
                    "Subcategory"
                ],
                "summary": "Get a list of subcategories for a parent category",
                "operationId": "a2537963b70f3d0cbd9972c5a3c00fd3",
                "parameters": [
                    {
                        "name": "parentId",
                        "in": "path",
                        "description": "Parent Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/subcategories": {
            "post": {
                "tags": [
                    "Subcategory"
                ],
                "summary": "Store a new subcategory",
                "operationId": "2f9457518c000f7b94d98f871c23beb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "parent_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/subcategories/{id}": {
            "get": {
                "tags": [
                    "Subcategory"
                ],
                "summary": "Get subcategory details",
                "operationId": "beb26b0b053dd3046453d1fbf9ff22b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subcategory not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subcategory"
                ],
                "summary": "Update a subcategory",
                "operationId": "ef5cf181180648682ffaaa5505acc916",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "parent_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mobile Development"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subcategory"
                ],
                "summary": "Delete a subcategory",
                "operationId": "d5683d77b74a03cfb5105d7703896769",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory deleted"
                    },
                    "404": {
                        "description": "Subcategory not found"
                    }
                }
            }
        },
...
    },
...
}